Add a way to view core commands in Sublime Merge.
This commit adds a way to view core commands in Merge as well. Since the
metadata folder is structured very similarly to Sublime Text, all we need
to do is configure the method to accept a parameter to decide whether
we want to view commands for one or the another.

The next commit will add the commands to the command palette.