[Partial Fix] Fix core command doc panel not opening for certain commands.
This commit introduces a partial fix, where the core command doc panel
was not opening for certain commands. Since the length padding can't be
negative in f-strings, we use `abs()` to get the absolute value. This
however, is not a perfect solution as all the `-` do not line up correctly
for such commands, but we can investigate that later.