[Fix] Use the `sublime.(find|load)_resources` APIs.
This commit fixes issues that cropped up because I had CommandsBrowser
in the packages directory. I was loading JSON files directly from the
folder, but when the package will be installed, that folder will go into
the installed packages and hence not available. So it can't read the files.

So, the above API usage will ensure that it works when both installed &
git cloned.